+++ packages/ingestor/src/lib/ingestLegacyObjectsToSkylark/main.ts
@@ -1,17 +1,8 @@
+/* eslint-disable no-console */
 import "../../env";
 import "./env";
 import { SAAS_API_ENDPOINT, SAAS_API_KEY } from "@skylark-reference-apps/lib";
-import {
-  ensureDir,
-  ensureFile,
-  exists,
-  outputFile,
-  readFile,
-  readJson,
-  readdir,
-  writeFile,
-} from "fs-extra";
-import { join } from "path";
+import { ensureDir } from "fs-extra";
 import { fetchObjectsFromLegacySkylark } from "./legacy";
 import {
   LegacyAsset,
@@ -19,7 +10,6 @@ import {
   LegacyCommonObject,
   LegacyEpisode,
   LegacyObjectType,
-  LegacyObjects,
   LegacySeason,
   LegacyTag,
   LegacyTagCategory,
@@ -37,21 +27,16 @@ import { setAccountConfiguration } from "../skylark/saas/account";
 import { createAlwaysAndForeverAvailability } from "../skylark/saas/availability";
 import { createObjectsInSkylark } from "./skylark";
 import { CreatedSkylarkObjects } from "./types/skylark";
-
-/* eslint-disable no-console */
-// For Macademia
-// - Miro: https://miro.com/app/board/uXjVMDOKvio=/
-
-// const objectTypes = [
-//   "Brand",
-//   "Season",
-//   "Episode",
-//   "Asset",
-//   "Availability", // Not in use / make everything always available (No dimensions)
-//   "Tags", // Game/App only
-//   "Imagery", // Ignore
-//   "Ratings", // Ignore
-// ];
+import {
+  createLegacyObjectsTimeStampedDir,
+  getMostRecentLegacyObjectsDir,
+  readObjectsFromFile,
+  writeAllLegacyObjectsToDisk,
+  writeError,
+  writeLegacyObjectsToDisk,
+  writeStatsForLegacyObjectsToDisk,
+} from "./fs";
+import { calculateTotalObjects } from "./utils";
 
 const checkEnvVars = () => {
   const legacyApiUrl = process.env.LEGACY_API_URL;
@@ -73,24 +58,6 @@ const checkEnvVars = () => {
   console.log("Skylark API URL:", SAAS_API_ENDPOINT);
 };
 
-const calculateTotalObjects = (
-  objects: Record<
-    string,
-    {
-      type: LegacyObjectType;
-      objects: Record<string, LegacyObjects>;
-      totalFound: number;
-    }
-  >
-) => {
-  const totalObjectsFound = Object.values(objects).reduce(
-    (previous, { totalFound }) => previous + totalFound,
-    0
-  );
-
-  return totalObjectsFound;
-};
-
 const updateSkylarkSchema = async ({
   assetTypes,
 }: {
@@ -109,119 +76,6 @@ const updateSkylarkSchema = async ({
     console.log("--- Activating Schema version:", updatedVersion);
     await activateConfigurationVersion(updatedVersion);
     await waitForUpdatingSchema(updatedVersion);
-  }
-};
-
-const generateEnvIdentifier = () => {
-  const onlyDataCreatedInLastMonth =
-    process.env.LEGACY_DATA_LAST_MONTH_ONLY === "true";
-
-  const env = process.env.LEGACY_API_URL
-    ? process.env.LEGACY_API_URL.replace("https://", "")
-        .replaceAll(".", "_")
-        .replaceAll("-", "_")
-    : "unknown";
-
-  if (onlyDataCreatedInLastMonth) {
-    return `${env}_last_month_only`;
-  }
-
-  return env;
-};
-
-const generateLegacyObjectsDir = () => {
-  const env = generateEnvIdentifier();
-
-  const dir = join(__dirname, "outputs", "legacy_data", env);
-  return dir;
-};
-
-const createLegacyObjectsTimeStampedDir = async () => {
-  const dateStamp = new Date().toISOString();
-
-  const dir = join(generateLegacyObjectsDir(), dateStamp);
-
-  await ensureDir(dir);
-
-  return dir;
-};
-
-const writeLegacyObjectsToDisk = async <T>(
-  dir: string,
-  obj: {
-    type: LegacyObjectType;
-    objects: Record<string, T[]>;
-    totalFound: number;
-  }
-) => {
-  try {
-    await writeFile(
-      join(dir, `${obj.type}.json`),
-      JSON.stringify(obj, null, 4)
-    );
-  } catch (err) {
-    console.error(
-      `[writeLegacyObjectsToDisk] Failed to write ${obj.type} data`
-    );
-  }
-};
-
-const writeAllLegacyObjectsToDisk = async (
-  objects: Record<
-    string,
-    {
-      type: LegacyObjectType;
-      objects: Record<string, LegacyObjects>;
-      totalFound: number;
-    }
-  >
-) => {
-  try {
-    const dir = await createLegacyObjectsTimeStampedDir();
-
-    await Promise.all(
-      Object.values(objects).map((value) =>
-        writeLegacyObjectsToDisk<LegacyObjects[0]>(dir, value)
-      )
-    );
-  } catch (err) {
-    console.error("[writeAllLegacyObjectsToDisk] Failed to write data");
-  }
-};
-
-const writeStatsForLegacyObjectsToDisk = async (
-  legacyObjects: Record<
-    string,
-    {
-      type: LegacyObjectType;
-      objects: Record<string, LegacyObjects>;
-      totalFound: number;
-    }
-  >
-) => {
-  try {
-    const env = generateEnvIdentifier();
-    const file = join(generateLegacyObjectsDir(), "stats.md");
-
-    await ensureFile(file);
-
-    const totalObjectsFound = calculateTotalObjects(legacyObjects);
-
-    let contents = `# Stats for ${env}\n\nTotal objects across ${USED_LANGUAGES.length} languages: ${totalObjectsFound}`;
-
-    Object.values(legacyObjects).forEach(({ type, objects, totalFound }) => {
-      contents += `\n\n## ${type} (${totalFound} total)\n`;
-
-      Object.entries(objects).forEach(([language, objs]) => {
-        contents += `  - \`${language.toLowerCase()}\`: \`${objs.length}\`\n`;
-      });
-
-      contents += "\n---";
-    });
-
-    await writeFile(file, contents);
-  } catch (err) {
-    console.error("[writeAllLegacyObjectsToDisk] Failed to write data");
   }
 };
 
@@ -284,41 +138,6 @@ const fetchLegacyObjects = async () => {
   );
 
   return retObj;
-};
-
-const getMostRecentLegacyObjectsDir = async () => {
-  const parentDir = generateLegacyObjectsDir();
-  await ensureDir(parentDir);
-
-  const objectDirs = (await readdir(parentDir)).filter(
-    (dir) => dir !== "stats.md"
-  );
-
-  const [mostRecentDir] = objectDirs.sort(
-    (a, b) => +new Date(b) - +new Date(a)
-  );
-
-  return join(parentDir, mostRecentDir);
-};
-
-const readObjectsFromFile = async <T>(dir: string, type: LegacyObjectType) => {
-  const file = join(dir, `${type}.json`);
-
-  if (!(await exists(file))) {
-    return {
-      type,
-      objects: {},
-      totalFound: 0,
-    };
-  }
-
-  const data = (await readJson(file)) as {
-    type: LegacyObjectType;
-    objects: Record<string, T[]>;
-    totalFound: number;
-  };
-
-  return data;
 };
 
 const readLegacyObjectsFromFile = async () => {
@@ -464,23 +283,6 @@ const main = async () => {
   );
 
   console.log("\nObjects Created Successfully.");
-};
-
-const writeError = async (err: unknown) => {
-  const errorFile = join(__dirname, "outputs", "ingest_errors.md");
-
-  await ensureFile(errorFile);
-
-  const contents = await readFile(errorFile, "utf8");
-
-  const dateStamp = new Date().toISOString();
-  const readFromDisk = process.env.READ_LEGACY_OBJECTS_FROM_DISK === "true";
-
-  const newContents = `${contents}\n---\n\n## ${dateStamp} (${
-    readFromDisk ? "Data from disk" : "Data from Skylark"
-  })\n\n\`\`\`\n${err as string}\n\`\`\`\n`;
-
-  await outputFile(errorFile, newContents);
 };
 
 main().catch((err) => {
